Extract report functions (2)
