+++ src/common/case-summary.report.ts
@@ -1,0 +1,5 @@
+import { Language } from "../ui";
+import { reportTextArea } from "../ui/text-area.report";
+
+export const reportCaseSummary = (value: string, language: Language) =>
+  reportTextArea("Renseignements cliniques", value, language);
