+++ src/ui/text-area.report.ts
@@ -1,0 +1,13 @@
+import { joinLines, Language, padSection, reportTitle } from ".";
+
+export const reportTextArea = (
+  title: string,
+  comments: string,
+  language: Language,
+) => {
+  const trimmedComments = comments.trim();
+
+  return trimmedComments
+    ? joinLines([reportTitle(title, language), padSection(trimmedComments)])
+    : undefined;
+};
